Globally exclude TestAssets
