+++ src/ZendCodingStandard/ruleset.xml
@@ -8,6 +8,9 @@
     <file>Sniffs/Files/COPYING.md</file>
     <file>Sniffs/Files/LICENSE.md</file>
 
+    <!-- Exclude test assets -->
+    <exclude-pattern>*/TestAsset/*</exclude-pattern>
+
     <!-- inherit rules from: -->
     <rule ref="PSR2"/>
     <rule ref="Generic.Arrays.DisallowLongArraySyntax"/>
