--- test/Sniffs/Commenting/Assets/FileLevelDocBlockMissing.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace ZendTest\CodingStandard\Sniffs\Commenting\Assets;
-
-class FileLevelDocBlockMissing
-{
-}
