+++ test/Sniffs/Commenting/TestAsset/FileLevelDocBlockIncorrectSourceLink.fixed.php
@@ -5,8 +5,8 @@
  * @license   https://github.com/zendframework/zend-coding-standard/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\CodingStandard\Sniffs\Commenting\Assets;
+namespace ZendTest\CodingStandard\Sniffs\Commenting\TestAsset;
 
-class FileLevelDocBlockIncorrectCopyrightLink
+class FileLevelDocBlockIncorrectSourceLink
 {
 }
