+++ test/Sniffs/Commenting/TestAsset/FileLevelDocBlockMissing.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace ZendTest\CodingStandard\Sniffs\Commenting\TestAsset;
+
+class FileLevelDocBlockMissing
+{
+}
