+++ test/Sniffs/Commenting/TestAsset/FileLevelDocBlockPass.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * @see       https://github.com/zendframework/zend-coding for the canonical source repository
+ * @see       https://github.com/zendframework/zend-coding-standard for the canonical source repository
  * @copyright https://github.com/zendframework/zend-coding-standard/blob/master/COPYING.md Copyright
  * @license   https://github.com/zendframework/zend-coding-standard/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\CodingStandard\Sniffs\Commenting\Assets;
+namespace ZendTest\CodingStandard\Sniffs\Commenting\TestAsset;
 
-class FileLevelDocBlockIncorrectSourceLink
+class FileLevelDocBlockPass
 {
 }
